refactor(sell-product): 统一账号缓存模型
