+++ tools/pipeline-generate/SellProduct/README.md
@@ -41,7 +41,7 @@ npx @joebao/maa-pipeline-generate --config pipeline-adb-config.json
 
 干员缓存按 UID 保存完整扫描快照。当前 UID 没有快照时，Pipeline 会先扫描 `operators` 中的全部干员，再开始售卖；存在快照时直接使用。启用强制刷新后，本次任务始终重新扫描完整干员表。只有首次建立或主动强制刷新的全局扫描可以写入缓存；据点内找人的局部扫描只服务于当次流程，不会覆盖已有快照。
 
-`SellProductOwnedOperators.json` 同时在独立的 `outpost_prosperity` 分区按 UID 保存据点发展值状态，不影响 `accounts` 中完整干员快照的判定。任务初始化时会加载上次确认的状态，使第一个据点开始规划时便能为后续未满据点预留发展值加成干员；未缓存的据点按未满处理。每次进入据点仍以右下角图标的最新识别结果为准，状态变化会立即更新当前会话和缓存，后续尚未完成的据点自动按新状态重新规划，已经完成的售后生产派驻保持锁定。
+`SellProductOwnedOperators.json` 在每个 UID 的账号分区中同时保存 `operators` 和 `locations`。`operators` 为 `null` 表示尚未完整扫描，空数组才表示已完整扫描但没有相关干员，因此只写入据点状态不会被误判为完整干员快照。任务初始化时会加载上次确认的 `locations`，使第一个据点开始规划时便能为后续未满据点预留发展值加成干员；未缓存的据点按未满处理。每次进入据点仍以右下角图标的最新识别结果为准，状态变化会立即更新当前会话和缓存，后续尚未完成的据点自动按新状态重新规划，已经完成的售后生产派驻保持锁定。
 
 物品保留规则提供 6 个独立槽位，每个槽位选择具体货品和最低保留数量。物品 case 通过 `attach` 提供 `item_id`，数量 input 通过 `custom_action_param.quantity` 提供整数。动态选品确认成功后记录实际 `itemId`，保留规则据此生效；数量 `0` 表示全部可售，同一物品重复配置时后面的槽位覆盖前面的槽位。
